+++ src/util/ark_query.py
@@ -16,6 +16,8 @@
 import requests
 import xmltodict
 
+from src import info
+
 
 def get_full_ark_url(prod_url):
     """
@@ -37,7 +39,7 @@ def quick_search(search_term):
     )
 
     try:
-        r = requests.get(url.format(search_term)).json()
+        r = requests.get(url.format(search_term), headers=info.useragent_header, timeout=info.requests_timeout).json()
         return r
     except Exception as e:
         if isinstance(e, requests.ConnectionError):
@@ -52,8 +54,9 @@ def get_codename(ark_url, tried=False):
     and obtain the HTML data.
     Parsing it, we can get the codename.
     """
+
     try:
-        text_thing = requests.get(ark_url).content.decode("utf8")
+        text_thing = requests.get(ark_url, headers=info.useragent_header, timeout=info.requests_timeout).content.decode("utf8")
     except Exception as e:
         if isinstance(e, requests.ConnectionError):
             return